creating listView API
