+++ src/controllers/listViewController.js
@@ -1,0 +1,9 @@
+import createStore from '../store';
+import listViewAPI from '../store/services/listView';
+
+const ListViewController = async (req, res) => {
+    const store = createStore();
+    await Promise.all(listViewAPI.util.getRunningOperationPromises());
+};
+
+export default ListViewController;
