+++ src/store/services/listView.js
@@ -1,0 +1,11 @@
+import { createApi } from '@reduxjs/toolkit/query/react';
+import graphqlBaseQuery from '../../services/gqlclient';
+
+const listViewAPI = createApi({
+    reducerPath: 'yelpApi',
+    baseQuery: graphqlBaseQuery,
+    endpoints: () => ({}),
+});
+
+export const { useGetBusinessQuery, useGetSearchQuery } = listViewAPI;
+export default listViewAPI;
