+++ src/store/services/yelp.js
@@ -1,32 +1,19 @@
 import { createApi } from '@reduxjs/toolkit/query/react';
-import { gql, ClientError } from 'graphql-request';
-import buildGQLClient from '../../services/gqlclient';
+import { gql } from 'graphql-request';
+import graphqlBaseQuery from '../../services/gqlclient';
 import getPostQuery from '../../query/getPost.graphql';
-
-const graphqlBaseQuery = async ({ body }) => {
-    try {
-        const graphQLClient = buildGQLClient();
-        const data = await graphQLClient.request(body);
-        return { data };
-    } catch (error) {
-        if (error instanceof ClientError) {
-            return { error: { status: error.response.status, data: error } };
-        }
-        return { error: { status: 500, data: error } };
-    }
-};
 
 const yelpAPI = createApi({
     reducerPath: 'yelpApi',
     baseQuery: graphqlBaseQuery,
     endpoints: (builder) => ({
         getBusiness: builder.query({
-            query: (id) => ({
+            query: () => ({
                 body: getPostQuery,
             }),
         }),
         getSearch: builder.query({
-            query: ({ term, location }) => ({
+            query: () => ({
                 body: gql`
                     query MyQuery {
                         posts {
